Minor changes + query.category

+++ backend/controllers/booksControllers.js
@@ -1,11 +1,13 @@
 import Book from '../models/Book.js'
 
+const projection = { createdAt: 0, updatedAt: 0, __v: 0 }
+
 const obtenerBooks = async (req, res) => {
   try {
     // booksByQuery
     if (req.query.name) {
       const { name } = req.query
-      const bookQuery = await Book.find({ 'nombre': { $regex: `^.*${name}.*` } },  '-createdAt -updatedAt -__v')
+      const bookQuery = await Book.find({ 'nombre': { $regex: `^.*${name}.*` } }, projection)
       if (bookQuery.length) {
         res.json(bookQuery)
       } else {
@@ -13,11 +15,22 @@ const obtenerBooks = async (req, res) => {
         res.status(404).json({ msg: error.message })
       }
 
+      // booksByCategory
+    } else if (req.query.category) {
+      const { category } = req.query
+      const categoryQuery = await Book.find({ category: { $in: [`${ category }`] } }, projection)
+      if (categoryQuery.length) {
+        res.json(categoryQuery)
+      } else {
+        const error = new Error('No se encontraron libros en esa categoría.')
+        res.status(404).json({ msg: error.message })
+      }
+
     } else {
       // getAllBooks      
-      const limit = req.query.limit  || 8
+      const limit = req.query.limit || 8
       const page = req.query.page || 1
-      const books = await Book.paginate({}, {limit, page})
+      const books = await Book.paginate({}, { projection, limit, page })
       const booksAll = await books.docs
 
       res.json(booksAll)
@@ -33,7 +46,7 @@ const nuevoBook = async (req, res) => {
 
   try {
     const bookAlmacenado = await book.save()
-    res.json(bookAlmacenado)
+    res.status(201).json(bookAlmacenado)
   } catch (error) {
     console.log(error)
   }
@@ -42,7 +55,7 @@ const nuevoBook = async (req, res) => {
 const detailBook = async (req, res) => {
   try {
     const { id } = req.params
-    const book = await Book.findById(id, '-createdAt -updatedAt -__v')
+    const book = await Book.findById(id, projection)
 
     if (!book) {
       const error = new Error('No se encontró el libro.')
@@ -60,46 +73,41 @@ const editarBook = async (req, res) => {
     const { id } = req.params
     const bookId = await Book.findById(id)
     if (bookId) {
-      const {
-        nombre,
-        descripcion,
-        colection,
-        category,
-        price,
-        rating
-      } = req.body
-
+      const { nombre, descripcion, colection, category, price, rating } = req.body
       await Book.updateOne({ nombre, descripcion, colection, category, price, rating })
     } else {
-      console.log('no hay libro con dicho ID')
+      const error = new Error('No se encontró el libro.')
+      res.status(404).json({ msg: error.message })
     }
-    res.send('Libro modificado')
-
+    res.json({ success_msg: 'Libro modificado correctamente' })
   } catch (error) {
     console.log(error)
   }
 }
 
 const eliminarBook = async (req, res) => {
-  try{
-    const {id} = req.params;
+  try {
+    const { id } = req.params
     const bookId = await Book.findById(id)
-    await Book.deleteOne({
-      where: {
-        id,
-      },
-    });
-    res.send('Libro eliminado correctamente')
-  } catch(error){
+    if (bookId) {
+      await Book.deleteOne({
+        where: { id },
+      })
+      res.json({ success_msg: 'Libro eliminado correctamente' })
+    } else {
+      const error = new Error('No se encontró el libro.')
+      res.status(404).json({ msg: error.message })
+    }
+  } catch (error) {
     console.log(error)
   }
 }
-
 
 export {
   obtenerBooks,
   detailBook,
   nuevoBook,
   editarBook,
-  eliminarBook
+  eliminarBook,
+  projection
 }
